CORDAOSM-1 dtp: remove complete_cb and timeout for dtp_reply_send
Remove both the complete_cb and timeout parameters for dtp_reply_send().
This could make the usage be simpler.

If the reply sending meets errors, either resend the reply internally
or just wait for the client to timeout and resend the request.
Currently handling is dtp does not resend the reply (only with an error
message).

Change-Id: Ib936b33ed3e7f83ff3021cc66b2b27b120471238
Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-on: http://review.whamcloud.com/18918
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Tested-by: Johann Lombardi <johann.lombardi@intel.com>