DAOS-14 vos: Object index implementation using DAOS btree
Definition and implementation of DAOS object internal
manipulation APIs. These are APIs which will be used by
the object cache layer which would be used by the object I/O
APIs. Implementation for object Index using daos btree.

This patch also introduces vos_init which needs to be called
for instantiation of a vos instance.

Change-Id: I142e88f23d1fdfae50dc193106a161ac3ebd176d
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: http://review.whamcloud.com/19703
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>