DAOS-82 build: allow successful build even if dependent components are broken
This is done be using components previously blessed by the newly
created daos-update-scratch job which runs on commits to master.

If scons install is specified, the build is put in PREFIX which
defaults to "install" can can be set on the command line.
If there is no scons local, prefix is set to #build to be completely
backwards compatible.

Change-Id: Ic331d9d5c8224dee575e28d1329f52d426bd1071
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: http://review.whamcloud.com/20038
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Tested-by: Johann Lombardi <johann.lombardi@intel.com>
Tested-by: Jenkins