DAOS-15 vos: VOS object cache implementation with an LRU DRAM cache
Implementation of VOS object cache using an LRU based DRAM.
This implementation uses a hash table in DRAM and a doubly-linked
list as the LRU cache. The hashtable is for O(1) lookup. The
doubly-linked list manages the records in the beginning or end
of the list, based on usage. (i.e) Most recently used is always on the
front and LRU are the ones pushed out. The cache interface provides
a way to access the objects within VOS located from object index table
in PMEM.

Change-Id: I1080f1179edfe5a0c32ee181317330dd0dafa87b
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: http://review.whamcloud.com/20020
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>