DAOS-83 server: simplify thread model
Move away from cgroups that turned out to be not very useful.
Use a simpler per-core binding based on hwloc which provides
a global HW topology that will be useful in the future.
The DAOS server can now be run as non-root.

This patch also stops service threads in parallel since this
is way faster.

Change-Id: I13717de90139a0cd0e629c177c0e64ec9db91fb3
Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>
Reviewed-on: http://review.whamcloud.com/20389
Tested-by: Jenkins