Add rpath to mercury
I used to be able to do -lcppr and pick up mercury and all of the
other dependencies automatically.   Now that we removed prerequisite
links, this no longer works.   But the root of the issue is that
paths are not saved in the mercury library for its dependencies.

With this change, I can do -lcppr just specifying the cppr path and
I can run only specifying the cppr path.

Change-Id: If4ca2a85e8a752ef3b5fe89ebde13cf1d7eba5dc
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: http://review.whamcloud.com/20958
Tested-by: Jenkins
Reviewed-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-by: Christopher Holguin <christopher.a.holguin@intel.com>