CORFSHIP-237 utest: Add unit test framework to CaRT
This change adds unit tests infrastructure to CaRT.  It also
adds a simple unit test for the purpose of checking the C/C++
linkage of CaRT header files.

Unit tests are compiled by default but to run them requires
the command line target utest.  This target can be modified
by --utest-mode={helgrind,memcheck} to run unit tests under
valgrind or helgrind.

The purpose of unit tests is for low level code to ensure it
is sound before building it into higher level components

I took this opportunity to play around with cmocka so the
one test that is added here shows some of the capabilities
of that interface.

Fix a few minor issues
1. Move the required callback to the top of
   dhash_table_ops_t to allow C++ users to initialize without
   setting preceeding fields to NULL
2. Remove C_PRINT to remove the temptation to clutter the
   console with messages in production code.
3. Fix memory leak caused by bad pointer check at shutdown

Change-Id: I4c7bc8f0edb82d5dcb97a55ded2decd33ff8e5ea
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: http://review.whamcloud.com/22799
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>