CART-37 build: CaRT builds have some issues
I've been seeing some strange issues with CaRT builds and
this change works around them.   For some reason, delaying
the built-in configure checks until a "require" call
causes some issues such as not recompiling after file
changes and very weird states with respect to configure
checks interferring with the actual builds where build
errors show up as configuration problems.

I have no idea why this happens but moving the
configuration step to the start (using "prebuild" kwarg)
seems to make all of these issues go away.  This option
only ensures that the prerequisites exist and doesn't
modify the environment.   It also ensures that all of
the command line options are available when we generate
cart-Linux.cfg.   We've been doing this in IOF and CPPR.

Actual changes made:
1. Uses prebuild kwarg for all required prerequisites
2. Adds a couple of missing headers to install
3. Adds uuid define to use scons_local configure check

So, without investigating further, this is a scons
and scons_local BKM.

Change-Id: Ia8b8a0eec98d16e32489341c4d87a30a2637b297
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: http://review.whamcloud.com/23704
Tested-by: Aditi Gaur <aditi.gaur@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>