DAOS-158 test: Replication test in degraded mode
Modified existing test code to independently invoke tests
Added a simple test to demo replication.

This test would perform a set of updates on different akey,
sleeps until a server is killed and a target is excluded (externally) and
continues to verify lookup of all updated keys in degraded mode,
sleeps further and continues to verify enumeration of all the
remaining keys.

Can use this test in to write regression tests for different kinds of failures that happen in the server, which can be introduced within the test instead of sleep
to verify correctness in degraded mode.

Demo for degraded mode can be invoked by "daos_test -d nkeys=1000,sleep=60".

Change-Id: I340c60ba012f6a83df8cbfb2df8c48b0dda702c6
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: http://review.whamcloud.com/23892
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>