DAOS-159 tests: Add new daosbench test "kv-simul"
This patch adds a new daosbench test "kv-simul", which mimics a
simulation that computes and stores multiple steps. Each step reads and
verifies the previous step's data, updates the data, and commit the
updates. If the test is killed in the middle, it can be restarted, with
exactly the same parameters. It then checks the last committed step and
resume from there.

Example:

  daosbench --test=kv-simul --testid=1 --dpool=$pool
	    --container=$container --pretty-print --aios=4 --keys=1000
	    --steps=3

Change-Id: Ie42a5521cea5a2aaedccf0b1ab200e431812028d
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/24026
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>