CART-100 build: Do not use prereq tools to import cart library.
The self-test code was using prereq tools to link in cart, which
did work but meant that libcrt itself was being checked for and
then built as a configure time check rather than the build phase.

This sets rpath to the install location of the code, so if
installed the code can run without LD_LIBRARY_PATH set.  If needing
to run from the build area then this will still be required.

Also remove some unneeded dependancies from elsewhere in the code.

Change-Id: I0e3754c4d1d016bcefc90078128a97ef3f5581b0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/24317
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>