DAOS-216 server: Add "--storage" option
This patch implements a hacky "--storage" option for specifying custom
storage path, which defaults to "/mnt/daos".

Assuming a dir path ends with "/" is a bit troublesome (incorrect IMHO).
This patch changes to assuming dir paths are well-formed (i.e., does not
end with "/") and always append a "/" before appending a new path
component. This works for both "/mnt/daos" and "/mnt/daos/".

Change-Id: I99998dc5bec96c80470aa31cc128a8f9c45fdc7b
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/24334
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>