CART-107 log: Allow log users to use the built-in mask
In order to have a single log in a process like the CNSS,
it is useful to have a single environment variable and call
crt_log_setmasks after a new batch of facilities is registered.
This change turns the error printed for unknown facilities
into a debug message since this is a commonly useful thing to
ignore.

With this change, a component can either call
crt_log_sync_mask
-or-
crt_log_setmasks

The former resets the mask after registering new facilities
using CRT_LOG_MASK.  The latter allows the user to override
the mask either manually or via a different environment variable

Change-Id: I5a6054a89c0222d378c4c5a824c6df69ffde338d
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/24410
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>