CART-133 usability: Restrict allowable characters for crt_group_id_t's
This patch restricts the characters that are allowable in crt_group_id_t
strings, and adds input validation to those public functions that accept
a crt_group_id_t as input. The allowable characters implemented in this
patch is the entire range of printable ASCII characters (between 0x20
and 0x7E, including space) except for quotation characters ('"`),
backslashes (\), and semicolons (;).

This enhances compatibility with the new self-test application - it is
now only possible to create groups for which self-test will be able to
test against.

Change-Id: Ib5314f2bea2d802e379d51cc139a6072ca108162
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/24506
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>