CART-124 self-test: Initial implementation of 1:many
Also addresses CART-123

This patch enhanced self-test with the ability to specify arbitrary
collections of ranks/tags to run self-test against. This allows testing
from one node against many nodes (1:many).

Self-test binds against a single group name (specified with --group-name)
and accepts one or more --endpoint arguments. Each endpoint is specified
as "ranks:tags". The ranks and tags are a comma-separated list of
integer identifiers (or a range of them). For example, "1-5,8:0-1"
specifies 6 ranks and two tags, resulting in 6*2 = 12 total endpoints
(equivalent to 1:0, 1:1, 2:0, 2:1, 3:0, 3:1, 4:0, 4:1, 5:0, 5:1, 8:0, 8:1).

Endpoints are executed sequentially in a round-robin fashion. Latency
measurements are now stamped internally with which rank/tag they were
measured against, though that information is currently not presented to
the user.

Change-Id: I765f474c7e779afefbde5e2dbe90553d55c3cb4a
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/24779
Tested-by: Maloo <hpdd-maloo@intel.com>