CART-125 self-test: Asymmetric message size configurability
This patch adds the ability to configure the self-test payload size in
both directions - the size sent to the service and the size of the
response. The --message-sizes argument to self-test is now a list of
tuples with the first element indicating the size to send to the
service, and the second indicated the reply size. For example (100, 200)
sends a 100-byte message to the service and receives a 200-byte
response.

This is implemented as four separate RPC opcodes which all
share the same general handler. They are used as follows:

(0 0) - Absolute minimum size message. No payload data is sent to the
	service and no payload is returned. This should be useful to make
	absolute minimum latency measurements without any argument marshalling.
(0, y) - A four-byte size integer is sent to the service which indicates
	that it should return a y-byte iovec
(x, 0) - An x-byte iovec is sent to the service which replies with an
	empty payload.
(x, y) - A four-byte size and an x-byte iovec is sent to the service,
	which replies with a y-byte iovec

Change-Id: I747758d99ff454106ee5c5a01bfb46fed3d67c38
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/24870
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>