CART-136 self-test: Improve self-test timeout/eviction robustness
Also addresses CART-126

This patch gives self-test the ability to properly detect and record RPC
errors. Errors returned to the self-test client by the RPC callback are
now recorded as a negative latency value which can be reported to the
user on a per-repetition basis (once reporting individual latencies is
implemented).

Self-test can now also handle eviction of nodes by CART. Evicted
endpoints(any endpoint that generates an error upon calling crt_req_send or
return -CER_OOG to the callback) are marked as evicted and no further attempts
are made to send messages to them.

The number of RPCs that fail is reported to the user, and only latencies
of RPCs which succeed are reported.

Change-Id: I34821cb7c1b969e572ce0bf50db66c5ac70c7b76
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/25067
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>