CQE-41 Add test list processing to multi runner
Enable to Multi Runner mode of Test Runner to process a list of test
description files. The list can be a list provided on the command line
or in a yaml file. The test list file can reference through a key in the
configuration file. The new key is called "test_list".

move CpprDaemonRunner to cppr

From 25249: Update the way that test logs are searched for.

When dumping logs after failure check for directories with names
matching anything after the first 'test_' rather than after the
second _ character.

Do not include the test_name in the loop directory name as it's
already used as the parent directory name.  This makes log paths
considerably shorter.

Change-Id: I4bdc2089fc9e3b648a4ce8baec4e98da9e64f0a6
Signed-off-by: mmennigx <martinx.mennig@intel.com>
Reviewed-on: https://review.whamcloud.com/25259
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>