CART-139 self-test: Add sessions to self-test
This changeset adds sessions to self-test. Sessions are utilized by
first sending an OPEN request to the service endpoint with the
parameters of the session (message sizes, etc). The service then returns
a session ID to use in the future when sending tests against that
session.

This provides a number of benefits relative to previous handling, the
biggest of which is that memory allocation is now not performed during
the actual test, and memory is able to be reclaimed when a session is
closed at the end of the test run.

Another benefit to this is a simple system to allocate a pool of buffers
to use for reply messages - previously, only one buffer was used for all
test RPCs. Each time a new test RPC arrives an available buffer is
popped off a stack and used to service the request before being re-added
to the stack of available buffers. This functionality will greatly
simplify the implementation of bulk messages in self test.

Also note that sessions are conveniently not required for test messages
that have a zero-byte reply.

Change-Id: I2701b2751f03495d519b14002c52163d93b62526
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/25366
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>