CART-131 self-test: Move self-test client into CART
This changeset integrates the self-test client component into CART. This
means that self-test sessions can now be run between any 1:many
CART-enabled ranks/tags, regardless of the application running on top.

The self-test command line client application sends the test parameters
to a "master" node (the '1' in 1:many) with a 'start' message containing
the parameters of the test session. That master node then opens
individual test sessions with each endpoint specified in the start
message, then runs the requested test RPC messages to evaluate
latency/ throughput / bandwidth, and finally returns all of the
resulting data to the command line application for display to the user.

Change-Id: If35c92eab30c8b803320b4e9339f54ff45f6d150
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/25742
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>