CART-120 self-test: Make test buffer memory alignment configurable
This changeset adds a new optional --align argument to self-test. This
can be used to force all test buffers to be aligned (or misaligned) as
specified.

This argument works by specifying what the least-significant byte of all
test buffer addresses should be forced to be. For example, if --align 0
is specified, all test buffer addresses will end in 0x00 (thus aligned
to 256 bytes).  To force misalignment, use something like --align 3. For
64-bit (8-byte) alignment, use something like --align 8 or --align 24
(0x08 and 0x18)

This changeset also fixes a session memory cleanup bug that was
discovered in crt_self_test_service.c

Change-Id: I9e4567be70d324a3237241aab1faac1907649536
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/26057
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>