DAOS-292 Patch to enable mocking of tests in scons_local
Patch searches the current directory for test_{mytest}.c files
and then searchs those files for UNIT_TEST entries. Files containing
unit tests are then added to the build along with the other files
required to create a working unit test.

This patch will require some changes to the build enviroment, in
particular mimick needs to be available. This update will be handled
else where in the process.

For a first step this patch

 - Detect test_{mytest}.c files and adds them to the build
 - Looks for UNIT_TEST s in the file and create a source file to be
   built with those tests in it
 - Bundles in other c files to allow a binary to be created
 - If a file called {mytest}.c exists in the parent then that file is
   also include.

Change-Id: Ie6ff3337e541ef4d905ae4be49eebf95b8221951
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: https://review.whamcloud.com/26101
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>