CART-174 self-test: Increase session_id to int64_t to fix struct packing
Prior to this patch, self-test used a 4-byte session_id. This session_id
is the first field in every self-test request that requires a session
(excludes EMPTY and SEND_IOV_REPLY_EMPTY messages). Unfortunately, using
a 4-byte session_id caused additional padding to be inserted into the
structures by the compiler. This was originally fixed by disabling
packing of these structs via #pragma pack, but this is not a good
long-term solution.

This changeset increases the session_id to 8 bytes to eliminate
misalignment and associated structure packing troubles.

Change-Id: I2ded43438b0693592736c884b408e536bc65fec3
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/26125
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>