CART-179 include: Implement crt_errstr()
This patch implements crt_errstr(), a utility function to turn the
values in crt_errno.h back into string values so they can be nicely
printed for the user.

This is implemented via some preprocessor trickery that in nevertheless
quite compact and eliminates the need to maintain the list of enum
constants in more than a single location. The way this works is that
everybody that includes crt_errno.h has the preprocessor definitions
reduce to a typedef enum { } block, but when crt_errno.c includes the
file, it reduces to a function with a  switch statement instead that
converts values back into their original string names.

Change-Id: I55ec79f15a387461c8d6915861c1fc14f65aa768
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/26195
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>