CART-130 self-test: Implement many:many
This changeset expands the --master-endpoint argument to accept a list
of endpoints that each act as a self-test origin running a 1:many test.
The self-test command line application is structured such that these
many 1:many tests happen in parallel (hence many:many).

The outermost loop in this application is over the message sizes. For
each message size, a 'start' message is sent to each master endpoint
with parameters for the 1:many test it should be responsible for
(message size, protocol, repetition count, etc.). Afterwards, the
command line application polls each master endpoint to see if it has
completed. Once all master endpoints have completed their tests the
results are printed out to the user.

Time measurements are still made entirely on the master (AKA origin)
nodes and reported to the command line application via the status
response - the poll interval has no effect on the accuracy of the test
time measurements.

Change-Id: I06505416e860c7d6d97ea44bd672b6c49c28e30e
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/26316
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>