DAOS-32 vos: Reporting completion of aggregation
Reporting and caching of scan completion of different levels within
vos_purge_anchor_t. Also added a flag to return the completion
of aggregation for use by layers above. This allows to optimize scanning
in scenarios of idempotent calls for aggregation.

Aggregation completion can be recorded in the VOS contaianer by saving
the aggregated epoch as a part of the vos_co_info_t object. Aggregation
completion reporting is currently triggered by passing an empty OID in
the call.

Change-Id: I7c89224c72af66b96da238db4fe57685b21b2688
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: https://review.whamcloud.com/26332
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>