CART-134 singleton: Modify singleton interface
This removes the environment variables to enable the
singleton usage.   The directory where the singleton file
is generated is controlled by the user and the creation of
the attach file is explicitly done via an API call rather
than automatically via an environment variable.

Furthermore, a service can now create an attach file for
any group to which it is attached.

This helps with two use cases

1. Since the application controls the location of the singleton
file, the application can ensure it is unique.   In CPPR and IOF, a
unique path must be given for the CNSS_PREFIX so this provides a
natural location where to store the singleton files.
2. For the interception library, I need a way to attach to the
IONSS without using PMIx.   This allows the CNSS to create a local
attach file on behalf of the IONSS which resides on another node.

Modify echo test so that the singleton and multi-tier modes can
be tested using command line options.

Change-Id: Ibade0a08998fa93012deb566052e7f879ecd44b3
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/26410
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>