CART-198 singleton: Emit only PSR to singleton attach file
This removes the need to lookup addresses for all members of
a set to generate the attach file as the size, psr rank, and
address are all that is needed.

Tested this change with the previously broken singleton mode
of the echo test with both multi tier and single tier and it
passes and generates the right information.

Change-Id: I1cb5e8f409bdc425490bd85cc7d6b647cda5a1c5
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/26603
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>