CART-188 self-test: Fix potential deadlock in self-test
Before this patch, there is a rare corner case for applications which
have only a single thread calling crt_progress(). If it just so happens
that all test buffers were in use (perhaps by ongoing BULK transfers),
self-test could possibly get into a state where it looped endlessly
trying to acquire a buffer that would never be made available (because
no other threads were calling crt_progress())

To avoid deadlocking, self-test will now print a warning message and
allocate an additional test buffer upon reaching a state that is likely
to deadlock. Even getting to this state is difficult to reproduce, so
test messages that are artificially delayed by having to perform
allocation should be rare.

Change-Id: I570e831e1928b16869bfa6f9d10092c83d40500b
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Reviewed-on: https://review.whamcloud.com/26645
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>