CQE-84 cppr results not being reported correctly
we are overwriting the top results file when Multi Runner executes more
than one description file. Each time Test Runner (or Multi Runner)
completes a description file it creates a subtest_results.yml file.
These files get collected at the end of the test job by the CI to create
the Maloo results.yml.

The file is being written to the top directory TestRun for each
description file Multi Runner is processing. It should be writing this
file in the top level directory for the description files. Which is a
directory under testRun.

Cleanup some log file naming.

move support for test cases to execute commands on other nodes to its
own class

Change-Id: I4b51751038ef1ae50fbb662d5b80d7d783881f27
Signed-off-by: mmennigx <martinx.mennig@intel.com>
Reviewed-on: https://review.whamcloud.com/27029
Tested-by: Jenkins
Reviewed-by: John Salinas <John.Salinas@intel.com>