DAOS-251: Tier Bugfixes and cleanup
--Initializing svc num struct (was left uninitialized)
--Fixed some memory leaks in failure paths in wireup
--Some small visual cleanup (old todos, etc)
--Small restructuring and edits of return codes for tier wireup
--Refactored cross connect test to use cmocka
--client side tier context is now manually set up
--added local tier information to server side

NOTE: ds_cross_conn_test expects two running tiers, with their
IDs handed in as arguments. Further, note tests expect a tier
that has not had any cross connect run on it, as it currently
does not clear server side state, so previous wire-ups will
cause the tests to fail as they expect particular return
codes that describe the state of the wireup.

Change-Id: I8b7e4e458aa8225b0547b3448b8ec4cf3dd1f5e6
Signed-off-by: Ian Adams <ian.f.adams@intel.com>
Reviewed-on: https://review.whamcloud.com/27331
Reviewed-by: John Keys <john.keys@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>