DAOS-280 common: Tolerate incomplete rsvc rank list
If users provide an incomplete rsvc rank list, for instance,

  actual:   1:2:3
  provided: 2

and current leader is 1, then an rsvc hint may contain rank 1, which
cannot be found in clients' rsvc rank list. We append the unknown rank
to the client rsvc rank list, instead of panicing, in this case.

Also, update the Jenkins build numbers for ompi, mercury, and cart to
bring them to sync with the commit SHAs.

Change-Id: I30568896fed86d819efcd1d8ebfb5b7630664a33
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/27487
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Jenkins