DAOS-260 object: lookup ds_cont directly
During object fetch/enumerate, if the cont hdl is
rebuild container hdl, let's lookup the ds_cont
directly, because there is only one rebuild container
hdl and might be used for mulitple container rebuild.

The patch is a bit hacky for now, but it will fix the
rebuild hang issue during my test.

Change-Id: Iedaae2dd19cacf2c492b6f44fa50b8ce8c1b9843
Signed-off-by: Wang Di <di.wang@intel.com>
Reviewed-on: https://review.whamcloud.com/27694
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>