CART-240 api: Change ep to pass-to-reference.
Change crt_req_send() and crt_ep_abort() to take a pointer to an
endpoint rather than passing a struct around on the stack.

This is an API change that will require modification of software layers
using CaRT.

Removes crt_idx and tgt_ep options from crt_hg_req_create() as they
were unsed.

Fixes CART-241, a crash on the target when an invalid rank is looked up

Change-Id: If1e7539bae6fe019f4420e99f9685d31adc5b867
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28122
Reviewed-by: Xuezhao Liu <xuezhao.liu@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>