CART-243 api: Add crt_req_set_endpoint() function.
Add a new function that can be used to modify a RPC endpoint after
it is created but before it is sent.  To use this new function a NULL
value must be passed to crt_req_create() and the request must then
be updated by a call to crt_req_set_endpoint() before it can be
injection into the network.

Change-Id: Ia4c5b9d311262f3c441a98c4c02b522d66ed022b
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28146
Reviewed-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>