DAOS-331: Add support to trigger NVML build in ARM
Detect ARM platform and launch ARM build for NVML.
This patch requires the NVML changes to be upstreamed.
Until that these changes are stored in a patch in
sl_patches directory as a patch for nvml while building
on ARM platforms. scons_local should automatically detect
and build in ARM platforms.

Change-Id: I4b04fdcf4a51a58095171193bcd9a01e89015bde
Signed-off-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-on: https://review.whamcloud.com/28507
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>