CART-89 test: Fix a bug and add documentation for python test code.
Fix a crash when testing for available sockets to assign to CaRT testing.
If the first socket selected was used then the code was crashing in
s.shutdown() before attempting the second loop.  Put in a assert to check
that after 10 repetitions a result has been found rather than continueing
with a known-bad value.

Change-Id: Ic7220cc9b3b1cb15d5e94db03fd13325fb529a94
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28401
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Martin Mennig <martinx.mennig@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>