CORCI-50 Update docker scripts for release jobs
fetch_dependent_artifacts.sh: Ignore mpi4py commit hash and
add support for *-release jobs.

The mpi4py matrix jobs are indexed by the ompi build they were built
using.  The cppr component needs an mpi4py built with the requested ompi.
The standalone builds build mpi4py as part of the ompi build and the
matrix jobs build mpi4py as a separate build.

The fetch_dependent_artifacts.sh script does an override of the wanted
mpi4py commit to use the ompi commit hash and this causes it to fail the
later check for a match.  Since mpi4py is currently obtained by the
scons-local web-retreiver in all cases, the commit hash in the
build.config is not used.

update_devel_branch.sh: Add support for -release and -master jobs.

Change-Id: Ib91b469d7930e6b15269f5627768f06525f9c20c
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>
Reviewed-on: https://review.whamcloud.com/28841
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>