CART-100 build: Check for supported -W compiler flags.
Maintain a list of known -W compiler options that should be used
if they are supported by the selected compiler.  Check at build
time which options are supported.

This allows the code to build with clang (if selected via a
sconslocalrc file) without setting any additional options.

The actual code in this changeset is a port from IOF.

Change-Id: I70851a4b3ea199b09186607b71f90dc647c30d38
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28943
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>