DAOS-320 rdb: Add term checking to rdb_tx_begin()
Add a term argument to rdb_tx_begin(), so that callers may explicitly
specify the term to begin a TX in, in order to make sure that the DB and
cache queries in the TX get consistent results. For example, the
following may happen in theory,

  /* Service is leader 8. */
  /* Somebody else becomes leader 9 and commits updates. */
  /* Raft replica becomes leader 10. */
  /* Service hasn't processed DOWN(8) and UP(10) yet. */
  rdb_tx_begin()	/* TX assigned to term 10 */
  rdb_tx_<query>()	/* TX reads from term 10 */
  <cache_query>()	/* TX reads from cache valid in term 8 */

The two queries may see inconsistent results. For simplicity, let's
begin all (except for the initialization TX) pool/container service TXs
in explicit terms, regardless of whether they query the cached pool map.

Change-Id: I0a11f2787556563d8e79b53dbf85aa77ca3fbee0
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.whamcloud.com/29244
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>