CART-89 errno: Modify gurt/errno.h to hide internals
The present definition of errno.h exposes the implementation
of an internal function.  Using a foreach allows for a more
efficient implementation of the internal function and doesn't
expose the details.  The range macro allows for multiple ranges
of contiguous error codes to be defined.

Aside from ensuring errno is always consistent, the header defined
this way allows for consistency in any future extentions that may
use these error codes such as a Python wrapper.   CPPR is a proof
of concept for this technique (with its own error codes).

Change-Id: I8b3f110afd04a2c4600adb457ff0e800be1ff288
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/29139
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-by: Yulu Jia <yulu.jia@intel.com>