CART-89 lm: fix double free of string field in RPC
fix a double free error. lm_uri_lookup_psr_cb did not make a copy of the
string output of the uri lookup RPC. This casued a double free becasue
CART frees the string when it destroyes the RPC, it also frees the same
string when it destroys the address cache.

Change-Id: I325e5fa3c0e402e7c7303c57076c5eaf5759a9da
Signed-off-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-on: https://review.whamcloud.com/29231
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>