CART-89 lm: fix double free of string field in RPC
fix a double free error. lm_uri_lookup_psr_cb did not make a copy of the
string output of the uri lookup RPC. This casued a double free becasue
CART frees the string when it destroyes the RPC, it also frees the same
string when it destroys the address cache.

This patch fixes the bug by stealing the memory pointed to by the RPC
output ptr, and setting the RPC output ptr to NULL. This is an (better)
alternative to 3510602d19f6be4530cf7f3ed98ce285fc76495e.

Change-Id: Idb2fd0adf076cb4767f13375a058f9ee626487f6
Signed-off-by: Yulu Jia <yulu.jia@intel.com>
Reviewed-on: https://review.whamcloud.com/29277
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>